--- packages/web/src/components/track-flair/Check.module.css
@@ -1,34 +1,0 @@
-.box.tiny {
-  height: 10px;
-  width: 10px;
-}
-
-.box.small {
-  height: 16px;
-  width: 16px;
-}
-
-.box.medium {
-  height: 24px;
-  width: 24px;
-}
-
-.box.large {
-  height: 32px;
-  width: 32px;
-}
-
-.box.xlarge {
-  height: 44px;
-  width: 44px;
-}
-
-.iconCoSign {
-  height: 100%;
-  width: 100%;
-}
-
-.iconCoSign rect {
-  stroke: var(--harmony-white);
-  fill: var(--harmony-static-primary);
-}
